# web/openapi.yaml
openapi: 3.1.0
info:
  title: SkillHub API
  description: |
    SkillHub registry HTTP API — discovery, publishing, authentication, namespaces,
    stars/comments/ratings, notifications, and admin operations.

    All JSON endpoints live under `/api/v1`. Bearer tokens (`Authorization: Bearer <token>`)
    are obtained via the device flow (`/api/v1/auth/device/*`) or via a logged-in web
    session cookie (`session_token`).

    See `/skills.md` for a runnable curl-based walkthrough designed for AI agents.
  version: "1.0.0"
  license:
    name: MIT
servers:
  - url: /
    description: Same-origin
...
paths:
...
  /api/v1/search:
    get:
      tags: [Search]
      summary: Full-text search over public skills
      parameters:
        - { name: q,      in: query, required: true,  schema: { type: string }, description: "Query string" }
        - { name: type,   in: query, required: false, schema: { type: string, enum: [skill, plugin, all], default: all }, description: "Filter by document type" }
        - { name: limit,  in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: offset, in: query, required: false, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: sort,   in: query, required: false, schema: { type: string }, description: "Bleve sort field, e.g. -downloads" }
      responses:
        "200":
          description: Bleve search result
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/skills:
    get:
      tags: [Skills]
      summary: List skills (paginated)
      parameters:
        - { name: cursor,   in: query, schema: { type: string },  description: "Opaque cursor from previous nextCursor" }
        - { name: sort,     in: query, schema: { type: string, enum: [created, downloads, stars, updated], default: created } }
        - { name: limit,    in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: category, in: query, schema: { type: string } }
      responses:
        "200":
          description: Paginated skill list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:       { type: array, items: { $ref: "#/components/schemas/Skill" } }
                  nextCursor: { type: string, nullable: true }
    post:
      tags: [Skills]
      summary: Publish a new skill version (multipart)
      description: |
        At least one file is required and `SKILL.md` must be present in the upload.
        With a team token, the `namespace` field MUST equal the token's bound namespace.
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [slug, version, files]
              properties:
                slug:         { type: string, description: "Unique skill slug" }
                version:      { type: string, description: "Semver, e.g. 0.1.0" }
                displayName:  { type: string }
                summary:      { type: string }
                category:     { type: string }
                kind:         { type: string }
                visibility:   { type: string, enum: [public, private] }
                namespace:    { type: string, description: "Namespace slug (required when using a team token)" }
                changelog:    { type: string }
                tags:         { type: string, description: "Comma- or semicolon-separated" }
                dependencies: { type: string, description: "JSON array of {slug, versionRange}" }
                signature:    { type: string, description: "Cosign bundle, max 64 KB" }
                files:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "200":
          description: Skill + new version
          content:
            application/json:
              schema:
                type: object
                properties:
                  skill:   { $ref: "#/components/schemas/Skill" }
                  version: { $ref: "#/components/schemas/SkillVersion" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/v1/download:
    get:
      tags: [Download]
      summary: Download a skill version as ZIP
      description: Honors `If-None-Match` for ETag-based revalidation.
      parameters:
        - { name: slug,    in: query, required: true,  schema: { type: string } }
        - { name: version, in: query, required: false, schema: { type: string, default: latest } }
        - { name: If-None-Match, in: header, required: false, schema: { type: string } }
      responses:
        "200":
          description: Skill bundle
          headers:
            ETag:                { schema: { type: string } }
            X-Skill-Version:     { schema: { type: string } }
            X-Skill-Fingerprint: { schema: { type: string } }
          content:
            application/zip:
              schema: { type: string, format: binary }
        "304": { description: Not Modified (ETag matched) }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/resolve:
    get:
      tags: [Download]
      summary: Reverse-lookup a version by SHA-256 fingerprint
      parameters:
        - { name: fingerprint, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Skill + version
          content:
            application/json:
              schema:
                type: object
                properties:
                  skill:   { $ref: "#/components/schemas/Skill" }
                  version: { $ref: "#/components/schemas/SkillVersion" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ──────────────────────────────────────────────────────────────────
  # Auth
  # ──────────────────────────────────────────────────────────────────
...
  /api/v1/whoami:
    get:
      tags: [Auth]
      summary: Inspect the currently authenticated principal
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ──────────────────────────────────────────────────────────────────
  # Personal tokens
  # ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Personal or team token from the device flow / `/api/v1/tokens`. Format is
        opaque; treat it as a secret string. Pass as `Authorization: Bearer <token>`.
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Set by `POST /login` and the OAuth callback. Browser-only.
  responses:
    BadRequest:
      description: Validation error.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Missing or invalid token.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Authenticated but not allowed.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource not found.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
...
    User:
      type: object
      properties:
        id:          { type: string, format: uuid }
        handle:      { type: string }
        displayName: { type: string, nullable: true }
        email:       { type: string, nullable: true }
        role:        { type: string, enum: [user, moderator, admin] }
        avatarUrl:   { type: string, nullable: true }
        createdAt:   { type: string, format: date-time }
    Skill:
      type: object
      properties:
        id:            { type: string, format: uuid }
        slug:          { type: string }
        displayName:   { type: string }
        summary:       { type: string }
        description:   { type: string }
        category:      { type: string, nullable: true }
        kind:          { type: string, nullable: true }
        tags:          { type: array, items: { type: string } }
        visibility:    { type: string, enum: [public, private] }
        latestVersion: { type: string }
        downloads:     { type: integer }
        stars:         { type: integer }
        ownerId:       { type: string, format: uuid }
        namespaceId:   { type: string, format: uuid, nullable: true }
        namespaceSlug: { type: string, description: "Slug of the namespace this skill belongs to" }
        createdAt:     { type: string, format: date-time }
        updatedAt:     { type: string, format: date-time }
    SkillVersion:
      type: object
      properties:
        id:              { type: string, format: uuid }
        skillId:         { type: string, format: uuid }
        version:         { type: string, description: "Semver" }
        changelog:       { type: string }
        fingerprint:     { type: string, description: "SHA-256 hex" }
        signatureBundle: { type: string, nullable: true }
        yankedAt:        { type: string, format: date-time, nullable: true }
        yankReason:      { type: string, nullable: true }
        deprecatedAt:    { type: string, format: date-time, nullable: true }
        deprecateMessage: { type: string, nullable: true }
        dependencies:
          type: array
          items:
            type: object
            properties:
              slug:         { type: string }
              versionRange: { type: string }
        publishedAt:     { type: string, format: date-time }
...